Rename all packages to follow the module name definition
