+++ examples/jaeger/src/main/java/io/opentelemetry/example/jaeger/JaegerExample.java
@@ -1,4 +1,4 @@
-package io.opentelemetry.example;
+package io.opentelemetry.example.jaeger;
 
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
